scribe/requirements.txt
numpy
SpeechRecognition
openai-whisper
torch
websockets
psutil
